stripe_example/requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
idna==3.3
MarkupSafe==2.1.1
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
tzdata==2022.2
urllib3==1.26.12
django-split-settings==1.2.0
django-dotenv==1.4.2
psycopg2==2.9.4
uwsgi==2.0.20
selenium==4.6.0
pydantic=1.10.2